+++ changelog.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Upcoming
+* Removed `iguanatweaksexpanded:iron_ores`, `gold_ores` and `copper_ores` block tags. These tags already exist in minecraft
+
 ## 1.12.1
 * Mobs now gain more Follow Range, XRay Range and movement speed the farther they spawn from world spawn
 * Beeg ore veins are now disabled by default